[15] Refactor : Vertical Scroll View를 재활용하기 위해 부모 클래스로 생성
